update abort implementation
