+++ api/assemblyscript/assembly/ffi.ts
@@ -1,5 +1,4 @@
 import { get_ffi_result } from "./env"
-import { fromFFI } from "./util"
 
 var current_ident: i32 = 0;
 
@@ -24,4 +23,28 @@ export function ffi_result(size: i32): ArrayBuffer {
 	}
 
 	return fromFFI(result_ptr, size)
+}
+
+export function fromFFI(ptr: usize, len: i32): ArrayBuffer {
+	let mem = new Uint8Array(len)
+
+	for (let i = 0; i < len; i++) {
+		mem[i] = load<u8>(ptr + i);
+	}
+
+	return mem.buffer
+}
+
+export class ffiValue {
+	ptr: usize
+	size: i32
+
+	constructor(ptr: usize, size: i32) {
+		this.ptr = ptr
+		this.size = size
+	}
+}
+
+export function toFFI(val: ArrayBuffer): ffiValue {
+	return new ffiValue(changetype<usize>(val), val.byteLength)
 }
