--- api/assemblyscript/index.js
@@ -1,5 +1,0 @@
-const fs = require("fs");
-const loader = require("@assemblyscript/loader");
-const imports = { /* imports go here */ };
-const wasmModule = loader.instantiateSync(fs.readFileSync(__dirname + "/build/optimized.wasm"), imports);
-module.exports = wasmModule.exports;
